requirements.txt
streamlit>=1.37
pymupdf>=1.24
sentence-transformers>=3.0
faiss-cpu>=1.8.0.post1
transformers>=4.44
torch>=2.2
numpy>=1.26
sentencepiece>=0.2
plotly>=5.18